# openapi/dist/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Reel Fake Api
  description: A REST api for buying or renting a movie dvd (only for educational purposes)
  license:
    name: MIT License
    url: https://mit-license.org
  contact:
    name: Pratap Reddy
    email: pratapreddy15@gmail.com
servers:
  - url: http://reelfake.cloud/api
    description: Production server
  - url: http://127.0.0.1:8000/api
    description: Local server
security:
  - cookieAuth: []
paths:
...
  /customers/{id}/forgot_password:
    put:
      summary: Change the customer password which is forgotten
      operationId: forgotCustomerPassword
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      requestBody:
        description: New password in the payload
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  format: password
                confirmedNewPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully changed the password
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: number
                  email:
                    type: string
                    format: email
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/{id}/change_password:
    put:
      summary: Change the customer password
      operationId: changeCustomerPassword
      security:
        - cookieAuth: []
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      requestBody:
        description: Payload for the current and new password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                confirmedNewPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully changed the password
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: number
                  email:
                    type: string
                    format: email
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers:
    get:
      summary: Get the customer list with pagination support
      operationId: getCustomers
      security:
        - cookieAuth: []
      tags:
        - customer
      parameters:
        - in: query
          name: page
          required: false
          description: |
            The page number (which starts with 1) to fetch the set of customers from.
            The customers are split into multiple pages with each page having a limited number of customers.
          schema:
            type: integer
            default: 1
        - in: query
          name: first_name
          required: false
          description: |
            The first name of the customer.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: first_name=%tap, first_name=pra% and first_name=%ata%.
          schema:
            type: string
        - in: query
          name: last_name
          required: false
          description: |
            The last name of the customers.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: last_name=%tap, last_name=pra% and last_name=%ata%.
          schema:
            type: string
        - in: query
          name: email
          required: false
          description: |
            The email of the customers.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: email=%tap, email=pra% and email=%ata%.
          schema:
            type: string
            format: email
        - in: query
          name: phone_number
          required: false
          description: |
            The phone number of the customer. This query supports only exact match.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: phone_number=%456, phone_number=123% and phone_number=%456%.
          schema:
            type: string
        - in: query
          name: registered_on
          required: false
          description: |
            Filters the customers by the registered date (YYYY-MM-DD). This can be a single registered date or a range. For e.g., ?registered_on=2024-01-01 or ?registered_on=2024-01-01,2024-06-30. The query string ?registered_on=2024-01-01 will query customers with the given registered date while ?registered_on=2024-01-01,2024-06-30 will query customers registered between 2024-01-01 and 2024-06-30.
          schema:
            type: string
            format: date
        - in: query
          name: address_line
          required: false
          description: |
            The address line of the customer resident.
            Filters the customer by the address line of the customer address.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: address_line=%street, address_line=any% and address_line=%str%.
          schema:
            type: string
        - in: query
          name: city
          required: false
          description: |
            The city of the customer resident.
            Filters the customer by the city of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: city=%abad, city=hyder% and city=%aba%.
          schema:
            type: string
        - in: query
          name: state
          required: false
          description: |
            The state of the customer resident.
            Filters the customer by the state of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: state=%gana, state=tel% and state=%anga%.
          schema:
            type: string
        - in: query
          name: country
          required: false
          description: |
            The country of the customer resident.
            Filters the customer by the country of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: country=%dia, country=ind% and country=%ndi%.
          schema:
            type: string
        - in: query
          name: postal_code
          required: false
          description: |
            The postal_code of the customer resident.
            Filters the customer by the postal_code of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: postal_code=%048, postal_code=500% and postal_code=%000%.
          schema:
            type: string
      responses:
        '200':
          description: Successfully returned list of customer summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/customerSummary'
                  length:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/pagination'
        '400':
          $ref: '#/components/responses/badRequest'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/summary:
    get:
      summary: Get the customer list with pagination support
      operationId: getCustomersSummary
      security:
        - cookieAuth: []
      tags:
        - customer
      parameters:
        - in: query
          name: page
          required: false
          description: |
            The page number (which starts with 1) to fetch the set of customers from.
            The customers are split into multiple pages with each page having a limited number of customers.
          schema:
            type: integer
            default: 1
        - in: query
          name: first_name
          required: false
          description: |
            The first name of the customer.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: first_name=%tap, first_name=pra% and first_name=%ata%.
          schema:
            type: string
        - in: query
          name: last_name
          required: false
          description: |
            The last name of the customers.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: last_name=%tap, last_name=pra% and last_name=%ata%.
          schema:
            type: string
        - in: query
          name: email
          required: false
          description: |
            The email of the customers.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: email=%tap, email=pra% and email=%ata%.
          schema:
            type: string
            format: email
        - in: query
          name: phone_number
          required: false
          description: |
            The phone number of the customer. This query supports only exact match.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: phone_number=%456, phone_number=123% and phone_number=%456%.
          schema:
            type: string
        - in: query
          name: registered_on
          required: false
          description: |
            Filters the customers by the registered date (YYYY-MM-DD). This can be a single registered date or a range. For e.g., ?registered_on=2024-01-01 or ?registered_on=2024-01-01,2024-06-30. The query string ?registered_on=2024-01-01 will query customers with the given registered date while ?registered_on=2024-01-01,2024-06-30 will query customers registered between 2024-01-01 and 2024-06-30.
          schema:
            type: string
            format: date
        - in: query
          name: address_line
          required: false
          description: |
            The address line of the customer resident.
            Filters the customer by the address line of the customer address.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: address_line=%street, address_line=any% and address_line=%str%.
          schema:
            type: string
        - in: query
          name: city
          required: false
          description: |
            The city of the customer resident.
            Filters the customer by the city of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: city=%abad, city=hyder% and city=%aba%.
          schema:
            type: string
        - in: query
          name: state
          required: false
          description: |
            The state of the customer resident.
            Filters the customer by the state of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: state=%gana, state=tel% and state=%anga%.
          schema:
            type: string
        - in: query
          name: country
          required: false
          description: |
            The country of the customer resident.
            Filters the customer by the country of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: country=%dia, country=ind% and country=%ndi%.
          schema:
            type: string
        - in: query
          name: postal_code
          required: false
          description: |
            The postal_code of the customer resident.
            Filters the customer by the postal_code of the customer resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: postal_code=%048, postal_code=500% and postal_code=%000%.
          schema:
            type: string
      responses:
        '200':
          description: Successfully returned list of customer summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                          format: email
                        active:
                          type: boolean
                  length:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/pagination'
        '400':
          $ref: '#/components/responses/badRequest'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/register:
    post:
      summary: Register a customer
      operationId: registerCustomer
      tags:
        - customer
      requestBody:
        description: Payload for the new customer
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: Successfully registered customer
          headers:
            Set-Cookie:
              description: The token generated from successful registration of customer
              schema:
                type: string
                example: auth_token=abcde12345
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/badRequest'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/{id}:
    get:
      summary: Get customer detail using the id
      operationId: getCustomerById
      security:
        - cookieAuth: []
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      responses:
        '200':
          description: Successfully returned the customer data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/customerDetail'
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
    put:
      summary: Update customer
      operationId: updateCustomer
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      requestBody:
        description: Payload of the movie to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                address:
                  type: object
                  properties:
                    addressLine:
                      type: string
                    cityName:
                      type: string
                    stateName:
                      type: string
                    country:
                      type: string
                    postalCode:
                      type: string
                phoneNumber:
                  type: string
                avatar:
                  type: string
                  format: uri
      responses:
        '204':
          description: Successfully updated the customer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
    delete:
      summary: |
        Delete the customer.
        Note: It is not allowed to delete a customer which is in use by any user
      operationId: deleteCustomer
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      responses:
        '204':
          description: Successfully deleted the customer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/{id}/activate:
    patch:
      summary: Activate customer
      operationId: activateCustomer
      security:
        - cookieAuth: []
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      responses:
        '204':
          description: Successfully updated the customer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/{id}/deactivate:
    patch:
      summary: Deactivate customer
      operationId: deactivateCustomer
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
      responses:
        '204':
          description: Successfully updated the customer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /customers/{id}/preferred_store/{store_id}:
    patch:
      summary: Set preferred store for the customer
      operationId: setPreferredStore
      tags:
        - customer
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the customer
          schema:
            type: integer
        - in: path
          name: store_id
          required: true
          description: The id of the store to be set as preferred on
          schema:
            type: integer
      responses:
        '204':
          description: Successfully updated the customer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: auth_token
      in: cookie
  schemas:
...
    pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Current page number
        totalPages:
          type: integer
          description: Total pages available for pagination
        totalItems:
          type: integer
          description: Total items across all the pages
        itemsPerPage:
          type: integer
          description: Maximum number of items that the page can have
        next:
          type: string
          description: Link to the next page
        prev:
          type: string
          description: Link to the previous page
        first:
          type: string
          description: Link to the first page
        last:
          type: string
          description: Link to the last page
...
    address:
      type: object
      properties:
        id:
          type: integer
          format: number
          readOnly: true
        addressLine:
          type: string
        cityName:
          type: string
        stateName:
          type: string
        country:
          type: string
        postalCode:
          type: string
    customerDetail:
      type: object
      properties:
        id:
          type: integer
          format: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
        phoneNumber:
          type: string
        avatar:
          type: string
          format: uri
        registeredOn:
          type: string
          format: date
        preferredStore:
          type: object
          properties:
            id:
              type: integer
              format: number
            phoneNumber:
              type: string
            storeManager:
              type: object
              properties:
                id:
                  type: string
                  format: number
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                active:
                  type: boolean
                phoneNumber:
                  type: string
                avatar:
                  type: string
                  format: uri
                address:
                  $ref: '#/components/schemas/address'
        address:
          $ref: '#/components/schemas/address'
...
    customerSummary:
      type: object
      properties:
        id:
          type: integer
          format: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        preferredStoreId:
          type: integer
          format: number
        active:
          type: boolean
        avatar:
          type: string
          format: uri
        registeredOn:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/address'
...
  responses:
    badRequest:
      description: The request does not have the data that api needed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    internalServerError:
      description: Server encountered unexpeted error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    unauthorized:
      description: Api key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
    notFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    forbidden:
      description: User do not have permission to perform the operation
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string